requirements/common.txt
django~=3.1.1
djangorestframework
djangorestframework-jwt
pillow
pilkit
django-imagekit